move chore into storagenode/contact package
