+++ storagenode/contact/chore.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2019 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package outreach
+package contact
 
 import (
 	"context"
@@ -10,7 +10,6 @@ import (
 
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
-	monkit "gopkg.in/spacemonkeygo/monkit.v2"
 
 	"storj.io/storj/internal/sync2"
 	"storj.io/storj/pkg/pb"
@@ -19,18 +18,14 @@ import (
 	"storj.io/storj/storagenode/trust"
 )
 
-var (
-	mon = monkit.Package()
-)
-
-// Config contains configurable parameters for outreach chore
+// Config contains configurable parameters for contact chore
 // TODO: Hide MaxSleep from CLI
 type Config struct {
-	Interval time.Duration `help:"how frequently the node outreach chore should run" releaseDefault:"1h" devDefault:"30s"`
+	Interval time.Duration `help:"how frequently the node contact chore should run" releaseDefault:"1h" devDefault:"30s"`
 	MaxSleep time.Duration `help:"maximum duration to wait before pinging satellites" releaseDefault:"45m" devDefault:"0s"`
 }
 
-// Chore is the outreach chore for nodes announcing themselves to their trusted satellites
+// Chore is the contact chore for nodes announcing themselves to their trusted satellites
 type Chore struct {
 	log       *zap.Logger
 	self      overlay.NodeDossier
@@ -42,7 +37,7 @@ type Chore struct {
 	Loop     *sync2.Cycle
 }
 
-// NewChore creates a new outreach chore
+// NewChore creates a new contact chore
 func NewChore(log *zap.Logger, interval time.Duration, maxSleep time.Duration, trust *trust.Pool, transport transport.Client, self overlay.NodeDossier) *Chore {
 	return &Chore{
 		log:       log,
@@ -56,10 +51,10 @@ func NewChore(log *zap.Logger, interval time.Duration, maxSleep time.Duration, t
 	}
 }
 
-// Run the outreach chore on a regular interval with jitter
+// Run the contact chore on a regular interval with jitter
 func (chore *Chore) Run(ctx context.Context) (err error) {
 	defer mon.Task()(&ctx)(&err)
-	chore.log.Info("Storagenode outreach chore starting up")
+	chore.log.Info("Storagenode contact chore starting up")
 
 	return chore.Loop.Run(ctx, func(ctx context.Context) error {
 		if err := chore.sleep(ctx); err != nil {
@@ -110,12 +105,6 @@ func (chore *Chore) pingSatellites(ctx context.Context) (err error) {
 	return group.Wait()
 }
 
-func awaitPingback() {
-	// TODO: write a method that listens for each satellite to ping back, or it times out and receives a log message
-	//  with which satellite they failed
-	//  make sure to close the connections regardless
-}
-
 // sleep for random interval in [0;maxSleep)
 // returns error if context was cancelled
 func (chore *Chore) sleep(ctx context.Context) error {
@@ -127,7 +116,7 @@ func (chore *Chore) sleep(ctx context.Context) error {
 	return nil
 }
 
-// Close stops the outreach chore
+// Close stops the contact chore
 func (chore *Chore) Close() error {
 	chore.Loop.Close()
 	return nil
