+++ storagenode/contact/chore_test.go
@@ -1,6 +1,6 @@
 // Copyright (C) 2019 Storj Labs, Inc.
 // See LICENSE for copying information.
 
-package outreach
+package contact
 
 //TODO
